Extract more to base-ui
