+++ app/src/main/kotlin/io/sweers/catchup/ui/about/LicensesFragment.kt
@@ -67,7 +67,7 @@ import io.sweers.catchup.ui.Scrollable
 import io.sweers.catchup.ui.StickyHeaders
 import io.sweers.catchup.ui.StickyHeadersLinearLayoutManager
 import io.sweers.catchup.ui.base.CatchUpItemViewHolder
-import io.sweers.catchup.ui.base.InjectableBaseFragment
+import io.sweers.catchup.base.ui.InjectableBaseFragment
 import io.sweers.catchup.util.UiUtil
 import io.sweers.catchup.util.dp2px
 import io.sweers.catchup.util.findSwatch
