+++ app/src/main/kotlin/io/sweers/catchup/ui/activity/OrderServicesActivity.kt
@@ -54,8 +54,8 @@ import io.sweers.catchup.injection.scopes.PerFragment
 import io.sweers.catchup.service.api.ServiceMeta
 import io.sweers.catchup.serviceregistry.ResolvedCatchUpServiceMetaRegistry
 import io.sweers.catchup.ui.FontHelper
-import io.sweers.catchup.ui.base.InjectableBaseFragment
-import io.sweers.catchup.ui.base.InjectingBaseActivity
+import io.sweers.catchup.base.ui.InjectableBaseFragment
+import io.sweers.catchup.base.ui.InjectingBaseActivity
 import io.sweers.catchup.base.ui.ColorUtils
 import io.sweers.catchup.util.asDayContext
 import io.sweers.catchup.util.isInNightMode
