+++ app/src/main/kotlin/io/sweers/catchup/ui/activity/ServiceSettingsActivity.kt
@@ -39,7 +39,7 @@ import io.sweers.catchup.service.api.ServiceConfiguration.ActivityConfiguration
 import io.sweers.catchup.service.api.ServiceConfiguration.PreferencesConfiguration
 import io.sweers.catchup.service.api.ServiceMeta
 import io.sweers.catchup.serviceregistry.ResolvedCatchUpServiceMetaRegistry
-import io.sweers.catchup.ui.base.InjectingBaseActivity
+import io.sweers.catchup.base.ui.InjectingBaseActivity
 import io.sweers.catchup.util.asDayContext
 import io.sweers.catchup.util.isInNightMode
 import io.sweers.catchup.util.setLightStatusBar
