+++ libraries/base-ui/src/main/kotlin/io/sweers/catchup/base/ui/CatchUpObjectWatcher.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2019 Zac Sweers
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.sweers.catchup.base.ui
+
+interface CatchUpObjectWatcher {
+  fun watch(watchedReference: Any)
+
+  object None : CatchUpObjectWatcher {
+    override fun watch(watchedReference: Any) {
+    }
+  }
+}
