+++ libraries/base-ui/src/main/kotlin/io/sweers/catchup/base/ui/InjectableBaseActivity.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2019. Zac Sweers
+ * Copyright (c) 2019 Zac Sweers
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,18 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.sweers.catchup.ui.base
+package io.sweers.catchup.base.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentManager.FragmentLifecycleCallbacks
 import dagger.android.AndroidInjection
-import io.sweers.catchup.app.CatchUpObjectWatcher
-import io.sweers.catchup.base.ui.BaseActivity
 import javax.inject.Inject
 
 abstract class InjectableBaseActivity : BaseActivity() {
+
+  @Inject
+  protected lateinit var viewContainer: ViewContainer
+  @Inject
+  protected lateinit var uiPreferences: UiPreferences
 
   override fun onCreate(savedInstanceState: Bundle?) {
     AndroidInjection.inject(this)
@@ -47,5 +50,10 @@ abstract class InjectableBaseActivity : BaseActivity() {
       }
     }
     supportFragmentManager.registerFragmentLifecycleCallbacks(callbacks, true)
+  }
+
+  override fun onAttachedToWindow() {
+    super.onAttachedToWindow()
+    updateNavBarColor(uiPreferences = uiPreferences)
   }
 }
