+++ libraries/base-ui/src/main/kotlin/io/sweers/catchup/base/ui/InjectingBaseFragment.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2019. Zac Sweers
+ * Copyright (c) 2019 Zac Sweers
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,14 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.sweers.catchup.ui.base
+package io.sweers.catchup.base.ui
 
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
 import dagger.android.HasAndroidInjector
 import javax.inject.Inject
 
-abstract class InjectingBaseActivity : InjectableBaseActivity(), HasAndroidInjector {
+abstract class InjectingBaseFragment : InjectableBaseFragment(), HasAndroidInjector {
 
   @Inject
   protected lateinit var androidInjector: DispatchingAndroidInjector<Any>
