+++ libraries/smmry/src/main/res/layout/fragment_smmry.xml
